% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Accordion}
\alias{Accordion}
\title{Accordion}
\usage{
Accordion(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/accordion/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item defaultExpanded \code{bool} \cr Default is FALSE If true, expands the accordion by default.
\item disabled \code{bool} \cr Default is FALSE If true, the component is disabled.
\item disableGutters \code{bool} \cr Default is FALSE If true, it removes the margin between two expanded accordion items and the increase of height.
\item expanded \code{bool} \cr Default is - If true, expands the accordion, otherwise collapse it. Setting this prop enables control over the accordion.
\item onChange \code{func} \cr Default is - Callback fired when the expand/collapse state is changed.Signature:function(event: React.SyntheticEvent, expanded: boolean) = voidevent The event source of the callback. Warning: This is a generic event not a change event.expanded The expanded state of the accordion.
\item slotProps \verb{\{ heading?: func| object, root?: func| object, transition?: func| object \}} \cr Default is The props used for each slot inside.
\item slots \verb{\{ heading?: elementType, root?: elementType, transition?: elementType \}} \cr Default is The components used for each slot inside.
\item square \code{bool} \cr Default is FALSE If true, rounded corners are disabled.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item TransitionComponent \code{elementType} \cr Default is - The component used for the transition. Follow this guide to learn more about the requirements for this component.Deprecated Use slots.transition instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item TransitionProps \code{object} \cr Default is - Props applied to the transition element. By default, the element is based on this Transition component.Deprecated Use slotProps.transition instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
}
}
\examples{
Accordion(
  AccordionSummary("Accordion 1"),
  AccordionDetails("Lorem ipsum dolor sit amet.")
)
}
