% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{AccordionSummary}
\alias{AccordionSummary}
\title{AccordionSummary}
\usage{
AccordionSummary(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/accordion-summary/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item expandIcon \code{node} \cr Default is - The icon to display as the expand indicator.
\item focusVisibleClassName \code{string} \cr Default is - This prop can help identify which element has keyboard focus. The class name will be applied when the element gains the focus through keyboard interaction. It's a polyfill for the CSS :focus-visible selector. The rationale for using this feature is explained here. A polyfill can be used to apply a focus-visible class to other components if needed.
\item slotProps \verb{\{ content?: func| object, expandIconWrapper?: func| object, root?: func| object \}} \cr Default is  The props used for each slot inside.
\item slots \verb{\{ content?: elementType, expandIconWrapper?: elementType, root?: elementType \}} \cr Default is  The components used for each slot inside.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
