% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Alert}
\alias{Alert}
\title{Alert}
\usage{
Alert(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/alert/}
}
\details{
\itemize{
\item action \code{node} \cr Default is - The action to display. It renders after the message, at the end of the alert.
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item closeText \code{string} \cr Default is 'Close' Override the default label for the close popup icon button.For localization purposes, you can use the provided translations.
\item color \code{'error'| 'info'| 'success'| 'warning'| string} \cr Default is - The color of the component. Unless provided, the value is taken from the severity prop. It supports both default and custom theme colors, which can be added as shown in the palette customization guide.
\item components \verb{\{ CloseButton?: elementType, CloseIcon?: elementType \}} \cr Default is The components used for each slot inside.Deprecated use the slots prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item componentsProps \verb{\{ closeButton?: object, closeIcon?: object \}} \cr Default is The extra props for the slot components. You can override the existing props or add new ones.Deprecated use the slotProps prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item icon \code{node} \cr Default is - Override the icon displayed before the children. Unless provided, the icon is mapped to the value of the severity prop. Set to false to remove the icon.
\item iconMapping \verb{\{ error?: node, info?: node, success?: node, warning?: node \}} \cr Default is - The component maps the severity prop to a range of different icons, for instance success to successoutlined. If you wish to change this mapping, you can provide your own. Alternatively, you can use the icon prop to override the icon displayed.
\item onClose \code{func} \cr Default is - Callback fired when the component requests to be closed. When provided and no action prop is set, a close icon button is displayed that triggers the callback when clicked.Signature:function(event: React.SyntheticEvent) = voidevent The event source of the callback.
\item role \code{string} \cr Default is 'alert' The ARIA role attribute of the element.
\item severity \code{'error'| 'info'| 'success'| 'warning'| string} \cr Default is 'success' The severity of the alert. This defines the color and icon used.
\item slotProps \verb{\{ action?: func| object, closeButton?: func| object, closeIcon?: func| object, icon?: func| object, message?: func| object, root?: func| object \}} \cr Default is The props used for each slot inside.
\item slots \verb{\{ action?: elementType, closeButton?: elementType, closeIcon?: elementType, icon?: elementType, message?: elementType, root?: elementType \}} \cr Default is The components used for each slot inside.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \code{'filled'| 'outlined'| 'standard'| string} \cr Default is 'standard' The variant to use.
}
}
\examples{
Alert(severity = "warning", AlertTitle("Warning"), "This is a warning Alert.")
}
