% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Avatar}
\alias{Avatar}
\title{Avatar}
\usage{
Avatar(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/avatar/}
}
\details{
\itemize{
\item alt \code{string} \cr Default is - Used in combination with src or srcSet to provide an alt attribute for the rendered img element.
\item children \code{node} \cr Default is - Used to render icon or text elements inside the Avatar if src is not set. This can be an element, or just a string.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item imgProps \code{object} \cr Default is - Attributes applied to the img element if the component is used to display an image. It can be used to listen for the loading error event.Deprecated Use slotProps.img instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item sizes \code{string} \cr Default is - The sizes attribute for the img element.
\item slotProps \verb{\{ img?: func| object \}} \cr Default is The props used for each slot inside.
\item slots \verb{\{ img?: elementType \}} \cr Default is The components used for each slot inside.
\item src \code{string} \cr Default is - The src attribute for the img element.
\item srcSet \code{string} \cr Default is - The srcSet attribute for the img element. Use this attribute for responsive image display.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \code{'circular'| 'rounded'| 'square'| string} \cr Default is 'circular' The shape of the avatar.
}
}
\examples{
Stack(
  direction = "row", spacing = 2,
  Avatar("H"), Avatar(sx = list(bgcolor = "secondary.main"), "N")
)
}
