% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{AvatarGroup}
\alias{AvatarGroup}
\title{AvatarGroup}
\usage{
AvatarGroup(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/avatar-group/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The avatars to stack.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item componentsProps \verb{\{ additionalAvatar?: object \}} \cr Default is - The extra props for the slot components. You can override the existing props or add new ones.This prop is an alias for the slotProps prop.Deprecated use the slotProps prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item max \code{number} \cr Default is 5 Max avatars to show before +x.
\item renderSurplus \code{func} \cr Default is - custom renderer of extraAvatarsSignature:function(surplus: number) =  React.ReactNodesurplus number of extra avatarsReturns: custom element to display
\item slotProps \verb{\{ additionalAvatar?: object, surplus?: func| object \}} \cr Default is  The props used for each slot inside.
\item slots \verb{\{ surplus?: elementType \}} \cr Default is  The components used for each slot inside.
\item spacing \code{'medium'| 'small'| number} \cr Default is 'medium' Spacing between avatars.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item total \code{number} \cr Default is children.length The total number of avatars. Used for calculating the number of extra avatars.
\item variant \code{'circular'| 'rounded'| 'square'| string} \cr Default is 'circular' The variant to use.
}
}
