% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{BottomNavigationAction}
\alias{BottomNavigationAction}
\title{BottomNavigationAction}
\usage{
BottomNavigationAction(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/bottom-navigation-action/}
}
\details{
\itemize{
\item children \code{unsupportedProp} \cr Default is NA This prop isn't supported. Use the component prop if you need to change the children structure.
\item classes \code{object} \cr Default is NA Override or extend the styles applied to the component.See CSS classes API below for more details.
\item icon \code{node} \cr Default is NA The icon to display.
\item label \code{node} \cr Default is NA The label element.
\item showLabel \code{bool} \cr Default is NA If true, the BottomNavigationAction will show its label. By default, only the selected BottomNavigationAction inside BottomNavigation will show its label.The prop defaults to the value (false) inherited from the parent BottomNavigation component.
\item sx \verb{Array func| object| bool | func| object} \cr Default is NA The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item value \code{any} \cr Default is NA You can provide your own value. Otherwise, we fallback to the child position index.
}
}
