% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Card}
\alias{Card}
\title{Card}
\usage{
Card(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/card/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item raised \code{bool} \cr Default is FALSE If true, the card will use raised styling.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
\examples{
Card(
  sx = list(maxWidth = 345),
  CardContent(
    Typography(variant = "h5", "Lizard"),
    Typography(variant = "body2", sx = list(color = "text.secondary"), "Lizards are reptiles.")
  ),
  CardActions(Button(size = "small", "Learn More"))
)
}
