% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{CardHeader}
\alias{CardHeader}
\title{CardHeader}
\usage{
CardHeader(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/card-header/}
}
\details{
\itemize{
\item action \code{node} \cr Default is - The action to display in the card header.
\item avatar \code{node} \cr Default is - The Avatar element to display.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item disableTypography \code{bool} \cr Default is FALSE If true, subheader and title won't be wrapped by a Typography component. This can be useful to render an alternative Typography variant by wrapping the title text, and optional subheader text with the Typography component.
\item slotProps \verb{\{ action?: func| object, avatar?: func| object, content?: func| object, root?: func| object, subheader?: func| object, title?: func| object \}} \cr Default is  The props used for each slot inside.
\item slots \verb{\{ action?: elementType, avatar?: elementType, content?: elementType, root?: elementType, subheader?: elementType, title?: elementType \}} \cr Default is  The components used for each slot inside.
\item subheader \code{node} \cr Default is - The content of the component.
\item subheaderTypographyProps \code{object} \cr Default is - These props will be forwarded to the subheader (as long as disableTypography is not true).Deprecated Use slotProps.subheader instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item title \code{node} \cr Default is - The content of the component.
\item titleTypographyProps \code{object} \cr Default is - These props will be forwarded to the title (as long as disableTypography is not true).Deprecated Use slotProps.title instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
}
}
