% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ClickAwayListener}
\alias{ClickAwayListener}
\title{ClickAwayListener}
\usage{
ClickAwayListener(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/click-away-listener/}
}
\details{
\itemize{
\item children \code{element} \cr Default is - The wrapped element.This needs to be able to hold a ref.
\item onClickAway \code{func} \cr Default is - Callback fired when a "click away" event is detected.
\item disableReactTree \code{bool} \cr Default is FALSE If true, the React tree is ignored and only the DOM tree is considered. This prop changes how portaled elements are handled.
\item mouseEvent \verb{'onClick'| 'onMouseDown'| 'onMouseUp'| 'onPointerDown'| 'onPointerUp'| false} \cr Default is 'onClick' The mouse event to listen to. You can disable the listener by providing false.
\item touchEvent \verb{'onTouchEnd'| 'onTouchStart'| false} \cr Default is 'onTouchEnd' The touch event to listen to. You can disable the listener by providing false.
}
}
