% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{Collapse}
\alias{Collapse}
\title{Collapse}
\usage{
Collapse(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/collapse/}
}
\details{
\itemize{
\item addEndListener \code{func} \cr Default is - Add a custom transition end trigger. Called with the transitioning DOM node and a done callback. Allows for more fine grained transition end logic. Note: Timeouts are still used as a fallback if provided.
\item children \code{node} \cr Default is - The content node to be collapsed.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item collapsedSize \code{number| string} \cr Default is '0px' The width (horizontal) or height (vertical) of the container when collapsed.
\item component \verb{element type} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.This needs to be able to hold a ref.
\item easing \verb{\{ enter?: string, exit?: string \}| string} \cr Default is - The transition timing function. You may specify a single easing or a object containing enter and exit values.
\item in \code{bool} \cr Default is - If true, the component will transition in.
\item orientation \code{'horizontal'| 'vertical'} \cr Default is 'vertical' The transition orientation.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item timeout \verb{'auto'| number| \{ appear?: number, enter?: number, exit?: number \}} \cr Default is duration.standard The duration for the transition, in milliseconds. You may specify a single timeout for all transitions, or individually with an object.Set to 'auto' to automatically calculate transition time based on height.
}
}
