% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Container}
\alias{Container}
\title{Container}
\usage{
Container(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/container/}
}
\details{
\itemize{
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item disableGutters \code{bool} \cr Default is FALSE If true, the left and right padding is removed.
\item fixed \code{bool} \cr Default is FALSE Set the max-width to match the min-width of the current breakpoint. This is useful if you'd prefer to design for a fixed set of sizes instead of trying to accommodate a fully fluid viewport. It's fluid by default.
\item maxWidth \code{'xs'| 'sm'| 'md'| 'lg'| 'xl'| false| string} \cr Default is 'lg' Determine the max-width of the container. The container width grows with the size of the screen. Set to false to disable maxWidth.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
\examples{
Container(maxWidth = "sm", Box(sx = list(bgcolor = "#cfe8fc", height = 200)))
}
