% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{CssBaseline}
\alias{CssBaseline}
\title{CssBaseline}
\usage{
CssBaseline(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/css-baseline/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - You can wrap a node.
\item enableColorScheme \code{bool} \cr Default is FALSE Enable color-scheme CSS property to use theme.palette.mode. For more details, check out https://developer.mozilla.org/en-US/docs/Web/CSS/color-scheme For browser support, check out https://caniuse.com/?search=color-scheme
}
}
