% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{FormControl}
\alias{FormControl}
\title{FormControl}
\usage{
FormControl(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/form-control/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item color \verb{'primary'| 'secondary'| 'error'| 'info'| 'success'| 'warning'| string} \cr Default is 'primary' The color of the component. It supports both default and custom theme colors, which can be added as shown in the palette customization guide.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item disabled \code{bool} \cr Default is FALSE If true, the label, input and helper text should be displayed in a disabled state.
\item error \code{bool} \cr Default is FALSE If true, the label is displayed in an error state.
\item focused \code{bool} \cr Default is - If true, the component is displayed in focused state.
\item fullWidth \code{bool} \cr Default is FALSE If true, the component will take up the full width of its container.
\item hiddenLabel \code{bool} \cr Default is FALSE If true, the label is hidden. This is used to increase density for a FilledInput. Be sure to add aria-label to the input element.
\item margin \verb{'dense'| 'none'| 'normal'} \cr Default is 'none' If dense or normal, will adjust vertical spacing of this and contained components.
\item required \code{bool} \cr Default is FALSE If true, the label will indicate that the input is required.
\item size \verb{'medium'| 'small'| string} \cr Default is 'medium' The size of the component.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \verb{'filled'| 'outlined'| 'standard'} \cr Default is 'outlined' The variant to use.
}
}
