% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{FormHelperText}
\alias{FormHelperText}
\title{FormHelperText}
\usage{
FormHelperText(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/form-helper-text/}
}
\details{
\itemize{
\item children \code{node} \cr Default is NA The content of the component.If ' ' is provided, the component reserves one line height for displaying a future message.
\item classes \code{object} \cr Default is NA Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is NA The component used for the root node. Either a string to use a HTML element or a component.
\item disabled \code{bool} \cr Default is NA If true, the helper text should be displayed in a disabled state.
\item error \code{bool} \cr Default is NA If true, helper text should be displayed in an error state.
\item filled \code{bool} \cr Default is NA If true, the helper text should use filled classes key.
\item focused \code{bool} \cr Default is NA If true, the helper text should use focused classes key.
\item margin \code{'dense'} \cr Default is NA If dense, will adjust vertical spacing. This is normally obtained via context from FormControl.
\item required \code{bool} \cr Default is NA If true, the helper text should use required classes key.
\item sx \verb{Array func| object| bool | func| object} \cr Default is NA The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \code{'filled'| 'outlined'| 'standard'| string} \cr Default is NA The variant to use.
}
}
