% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{FormLabel}
\alias{FormLabel}
\title{FormLabel}
\usage{
FormLabel(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/form-label/}
}
\details{
\itemize{
\item children \code{node} \cr Default is NA The content of the component.
\item classes \code{object} \cr Default is NA Override or extend the styles applied to the component.See CSS classes API below for more details.
\item color \verb{'error'| 'info'| 'primary'| 'secondary'| 'success'| 'warning'| string} \cr Default is NA The color of the component. It supports both default and custom theme colors, which can be added as shown in the palette customization guide.
\item component \code{elementType} \cr Default is NA The component used for the root node. Either a string to use a HTML element or a component.
\item disabled \code{bool} \cr Default is NA If true, the label should be displayed in a disabled state.
\item error \code{bool} \cr Default is NA If true, the label is displayed in an error state.
\item filled \code{bool} \cr Default is NA If true, the label should use filled classes key.
\item focused \code{bool} \cr Default is NA If true, the input of this label is focused (used by FormGroup components).
\item required \code{bool} \cr Default is NA If true, the label will indicate that the input is required.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is NA The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
