% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{GlobalStyles}
\alias{GlobalStyles}
\title{GlobalStyles}
\usage{
GlobalStyles(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/global-styles/}
}
\details{
\itemize{
\item styles \code{array| func| number| object| string| bool} \cr Default is NA The styles you want to apply globally.
}
}
