% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{GridLegacy}
\alias{GridLegacy}
\title{GridLegacy}
\usage{
GridLegacy(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/grid-legacy/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item columns \verb{Array number | number| object} \cr Default is 12 The number of columns.
\item columnSpacing \verb{Array number| string | number| object| string} \cr Default is - Defines the horizontal space between the type item components. It overrides the value of the spacing prop.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item container \code{bool} \cr Default is FALSE If true, the component will have the flex container behavior. You should be wrapping items with a container.
\item direction \verb{'column-reverse'| 'column'| 'row-reverse'| 'row'| Array 'column-reverse'| 'column'| 'row-reverse'| 'row' | object} \cr Default is 'row' Defines the flex-direction style property. It is applied for all screen sizes.
\item item \code{bool} \cr Default is FALSE If true, the component will have the flex item behavior. You should be wrapping items with a container.
\item lg \code{'auto'| number| bool} \cr Default is FALSE If a number, it sets the number of columns the grid item uses. It can't be greater than the total number of columns of the container (12 by default). If 'auto', the grid item's width matches its content. If false, the prop is ignored. If true, the grid item's width grows to use the space available in the grid container. The value is applied for the lg breakpoint and wider screens if not overridden.
\item md \code{'auto'| number| bool} \cr Default is FALSE If a number, it sets the number of columns the grid item uses. It can't be greater than the total number of columns of the container (12 by default). If 'auto', the grid item's width matches its content. If false, the prop is ignored. If true, the grid item's width grows to use the space available in the grid container. The value is applied for the md breakpoint and wider screens if not overridden.
\item rowSpacing \verb{Array number| string | number| object| string} \cr Default is - Defines the vertical space between the type item components. It overrides the value of the spacing prop.
\item sm \code{'auto'| number| bool} \cr Default is FALSE If a number, it sets the number of columns the grid item uses. It can't be greater than the total number of columns of the container (12 by default). If 'auto', the grid item's width matches its content. If false, the prop is ignored. If true, the grid item's width grows to use the space available in the grid container. The value is applied for the sm breakpoint and wider screens if not overridden.
\item spacing \verb{Array number| string | number| object| string} \cr Default is 0 Defines the space between the type item components. It can only be used on a type container component.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item wrap \code{'nowrap'| 'wrap-reverse'| 'wrap'} \cr Default is 'wrap' Defines the flex-wrap style property. It's applied for all screen sizes.
\item xl \code{'auto'| number| bool} \cr Default is FALSE If a number, it sets the number of columns the grid item uses. It can't be greater than the total number of columns of the container (12 by default). If 'auto', the grid item's width matches its content. If false, the prop is ignored. If true, the grid item's width grows to use the space available in the grid container. The value is applied for the xl breakpoint and wider screens if not overridden.
\item xs \code{'auto'| number| bool} \cr Default is FALSE If a number, it sets the number of columns the grid item uses. It can't be greater than the total number of columns of the container (12 by default). If 'auto', the grid item's width matches its content. If false, the prop is ignored. If true, the grid item's width grows to use the space available in the grid container. The value is applied for all the screen sizes with the lowest priority.
\item zeroMinWidth \code{bool} \cr Default is FALSE If true, it sets min-width: 0 on the item. Refer to the limitations section of the documentation to better understand the use case.
}
}
