% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ImageList}
\alias{ImageList}
\title{ImageList}
\usage{
ImageList(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/image-list/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component, normally ImageListItems.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item cols \code{integer} \cr Default is 2 Number of columns.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item gap \code{number} \cr Default is 4 The gap between items in px.
\item rowHeight \code{'auto'| number} \cr Default is 'auto' The height of one row in px.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \code{'masonry'| 'quilted'| 'standard'| 'woven'| string} \cr Default is 'standard' The variant to use.
}
}
