% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ImageListItem}
\alias{ImageListItem}
\title{ImageListItem}
\usage{
ImageListItem(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/image-list-item/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component, normally an \if{html}{\out{<img>}}.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item cols \code{integer} \cr Default is 1 Width of the item in number of grid columns.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item rows \code{integer} \cr Default is 1 Height of the item in number of grid rows.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
