% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{knit_print.muiMaterial}
\alias{knit_print.muiMaterial}
\title{Knit print method for muiMaterial components}
\usage{
\method{knit_print}{muiMaterial}(x, ...)
}
\arguments{
\item{x}{A muiMaterial object.}

\item{...}{Additional arguments passed to the underlying method.}
}
\value{
knitr output suitable for HTML documents.
}
\description{
Renders muiMaterial components as HTML in knitr/Quarto documents.
}
