% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{List}
\alias{List}
\title{List}
\usage{
List(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/list/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item dense \code{bool} \cr Default is FALSE If true, compact vertical padding designed for keyboard and mouse input is used for the list and list items. The prop is available to descendant components as the dense context.
\item disablePadding \code{bool} \cr Default is FALSE If true, vertical padding is removed from the list.
\item subheader \code{node} \cr Default is - The content of the subheader, normally ListSubheader.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
\examples{
List(
  ListItem(disablePadding = TRUE, ListItemButton(ListItemText(primary = "Inbox"))),
  ListItem(disablePadding = TRUE, ListItemButton(ListItemText(primary = "Drafts")))
)
}
