% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ListItem}
\alias{ListItem}
\title{ListItem}
\usage{
ListItem(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/list-item/}
}
\details{
\itemize{
\item alignItems \verb{'center'| 'flex-start'} \cr Default is 'center' Defines the align-items style property.
\item children \code{node} \cr Default is - The content of the component if a ListItemSecondaryAction is used it must be the last child.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item components \verb{\{ Root?: elementType \}} \cr Default is  The components used for each slot inside.Deprecated－Use the slots prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item componentsProps \verb{\{ root?: object \}} \cr Default is  The extra props for the slot components. You can override the existing props or add new ones.Deprecated－Use the slotProps prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item ContainerComponent \verb{element type} \cr Default is 'li' The container component used when a ListItemSecondaryAction is the last child.This needs to be able to hold a ref.Deprecated－Use the component or slots.root prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item ContainerProps \code{object} \cr Default is  Props applied to the container component if used.Deprecated－Use the slotProps.root prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item dense \code{bool} \cr Default is FALSE If true, compact vertical padding designed for keyboard and mouse input is used. The prop defaults to the value inherited from the parent List component.
\item disableGutters \code{bool} \cr Default is FALSE If true, the left and right padding is removed.
\item disablePadding \code{bool} \cr Default is FALSE If true, all padding is removed.
\item divider \code{bool} \cr Default is FALSE If true, a 1px light border is added to the bottom of the list item.
\item secondaryAction \code{node} \cr Default is - The element to display at the end of ListItem.
\item slotProps \verb{\{ root?: object \}} \cr Default is  The extra props for the slot components. You can override the existing props or add new ones.
\item slots \verb{\{ root?: elementType \}} \cr Default is  The components used for each slot inside.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
