% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/inputs.R
\name{ListItemButton}
\alias{ListItemButton}
\alias{ListItemButton.shinyInput}
\alias{updateListItemButton.shinyInput}
\title{ListItemButton}
\usage{
ListItemButton(...)

ListItemButton.shinyInput(inputId, ...)

updateListItemButton.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to the component.}

\item{inputId}{ID of the component.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/list-item-button/}
}
\details{
\itemize{
\item alignItems \code{'center'| 'flex-start'} \cr Default is 'center' Defines the align-items style property.
\item autoFocus \code{bool} \cr Default is FALSE If true, the list item is focused during the first mount. Focus will also be triggered if the value changes from false to true.
\item children \code{node} \cr Default is - The content of the component if a ListItemSecondaryAction is used it must be the last child.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item dense \code{bool} \cr Default is FALSE If true, compact vertical padding designed for keyboard and mouse input is used. The prop defaults to the value inherited from the parent List component.
\item disabled \code{bool} \cr Default is FALSE If true, the component is disabled.
\item disableGutters \code{bool} \cr Default is FALSE If true, the left and right padding is removed.
\item divider \code{bool} \cr Default is FALSE If true, a 1px light border is added to the bottom of the list item.
\item focusVisibleClassName \code{string} \cr Default is - This prop can help identify which element has keyboard focus. The class name will be applied when the element gains the focus through keyboard interaction. It's a polyfill for the CSS :focus-visible selector. The rationale for using this feature is explained here. A polyfill can be used to apply a focus-visible class to other components if needed.
\item selected \code{bool} \cr Default is FALSE Use to apply selected styling.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
\note{
\code{input[[inputId]]} is a click counter (like
  \code{shiny::actionButton()}), incremented each time the row is clicked.
  For lists generated dynamically with \code{lapply()}, an alternative to
  one input per row is a single custom event carrying the clicked row as
  payload: \code{onClick = JS("() => Shiny.setInputValue('rowClicked',
  'row-1', {priority: 'event'})")}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
library(shiny)
library(muiMaterial)

ui <- muiMaterialPage(
  List(
    ListItemButton.shinyInput(
      inputId = "item1",
      ListItemText(primary = "Click me")
    )
  ),
  verbatimTextOutput("count")
)

server <- function(input, output, session) {
  output$count <- renderPrint(input$item1)
}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
