% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ListItemIcon}
\alias{ListItemIcon}
\title{ListItemIcon}
\usage{
ListItemIcon(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/list-item-icon/}
}
\details{
\itemize{
\item children \code{node} \cr Default is NA The content of the component, normally Icon, SvgIcon, or a @mui/icons-material SVG icon element.
\item classes \code{object} \cr Default is NA Override or extend the styles applied to the component.See CSS classes API below for more details.
\item sx \verb{Array func| object| bool | func| object} \cr Default is NA The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
