% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ListItemText}
\alias{ListItemText}
\title{ListItemText}
\usage{
ListItemText(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/list-item-text/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - Alias for the primary prop.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item disableTypography \code{bool} \cr Default is FALSE If true, the children won't be wrapped by a Typography component. This can be useful to render an alternative Typography variant by wrapping the children (or primary) text, and optional secondary text with the Typography component.
\item inset \code{bool} \cr Default is FALSE If true, the children are indented. This should be used if there is no left avatar or left icon.
\item primary \code{node} \cr Default is - The main content element.
\item primaryTypographyProps \code{object} \cr Default is - These props will be forwarded to the primary typography component (as long as disableTypography is not true).Deprecated Use slotProps.primary instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item secondary \code{node} \cr Default is - The secondary content element.
\item secondaryTypographyProps \code{object} \cr Default is - These props will be forwarded to the secondary typography component (as long as disableTypography is not true).Deprecated Use slotProps.secondary instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item slotProps \verb{\{ primary?: func| object, root?: func| object, secondary?: func| object \}} \cr Default is  The props used for each slot inside.
\item slots \verb{\{ primary?: elementType, root?: elementType, secondary?: elementType \}} \cr Default is  The components used for each slot inside.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
