% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muiMaterialDependency.R
\name{muiMaterialDependency}
\alias{muiMaterialDependency}
\title{muiMaterial JS dependency}
\usage{
muiMaterialDependency()
}
\value{
HTML dependency object.
}
\description{
muiMaterial JS dependency
}
