% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{NoSsr}
\alias{NoSsr}
\title{NoSsr}
\usage{
NoSsr(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/no-ssr/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - You can wrap a node.
\item defer \code{bool} \cr Default is FALSE If true, the component will not only prevent server-side rendering. It will also defer the rendering of the children into a different screen frame.
\item fallback \code{node} \cr Default is null The fallback content to display.
}
}
