% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Paper}
\alias{Paper}
\title{Paper}
\usage{
Paper(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/paper/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item elevation \code{integer} \cr Default is 1 Shadow depth, corresponds to dp in the spec. It accepts values between 0 and 24 inclusive.
\item square \code{bool} \cr Default is FALSE If true, rounded corners are disabled.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \code{'elevation'| 'outlined'| string} \cr Default is 'elevation' The variant to use.
}
}
\examples{
Stack(
  direction = "row", spacing = 2,
  Paper(elevation = 0, "0"), Paper("1"), Paper(elevation = 3, "3")
)
}
