% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{PigmentStack}
\alias{PigmentStack}
\title{PigmentStack}
\usage{
PigmentStack(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/pigment-stack/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item direction \verb{'column-reverse'| 'column'| 'row-reverse'| 'row'| Array 'column-reverse'| 'column'| 'row-reverse'| 'row' | \{ lg?: 'column-reverse'| 'column'| 'row-reverse'| 'row', md?: 'column-reverse'| 'column'| 'row-reverse'| 'row', sm?: 'column-reverse'| 'column'| 'row-reverse'| 'row', xl?: 'column-reverse'| 'column'| 'row-reverse'| 'row', xs?: 'column-reverse'| 'column'| 'row-reverse'| 'row' \}} \cr Default is 'column' Defines the flex-direction style property. It is applied for all screen sizes.
\item divider \code{node} \cr Default is - Add an element between each child.
\item spacing \verb{Array number| string | number| \{ lg?: number| string, md?: number| string, sm?: number| string, xl?: number| string, xs?: number| string \}| string} \cr Default is 0 Defines the space between immediate children.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop, which allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
