% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{runExample}
\alias{runExample}
\title{Run shinyMaterialUI example}
\usage{
runExample(name)
}
\arguments{
\item{name}{Use \link{getExamples} to get
the available examples.}
}
\value{
Runs a Shiny app.
}
\description{
Run shinyMaterialUI example
}
