% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{shinyMaterialUIPage}
\alias{shinyMaterialUIPage}
\title{HTML body without Bootstrap and margins}
\usage{
shinyMaterialUIPage(..., suppressBootstrap = TRUE, debug_react = FALSE)
}
\arguments{
\item{...}{The contents of the document body.}

\item{suppressBootstrap}{Whether to suppress Bootstrap.}

\item{debug_react}{Whether to enable react debug mode. Default to FALSE.}
}
\value{
html object with 'margin:0' which can be passed as the UI of a Shiny app.
}
\description{
Creates a Material UI page without using CssBaseline, so
you can use regular Shiny inputs or shiny.router.
}
\details{
The Bootstrap library is suppressed by default, as it doesn't work well
with Material UI in general. Some default Material UI styles are apply, i.e. the
margin is removed for all browser and the font size.


\url{https://mui.com/material-ui/react-css-baseline/}
}
