% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Skeleton}
\alias{Skeleton}
\title{Skeleton}
\usage{
Skeleton(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/skeleton/}
}
\details{
\itemize{
\item animation \code{'pulse'| 'wave'| false} \cr Default is 'pulse' The animation. If false the animation effect is disabled.
\item children \code{node} \cr Default is - Optional children to infer width and height from.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item height \code{number| string} \cr Default is - Height of the skeleton. Useful when you don't want to adapt the skeleton to a text element but for instance a card.
\item sx \code{Array} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \code{'circular'| 'rectangular'| 'rounded'| 'text'| string} \cr Default is 'text' The type of content that will be rendered.
\item width \code{number| string} \cr Default is - Width of the skeleton. Useful when the skeleton is inside an inline element with no width of its own.
}
}
\examples{
Stack(
  spacing = 1,
  Skeleton(variant = "text"),
  Skeleton(variant = "rectangular", width = 210, height = 60)
)
}
