% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{SnackbarContent}
\alias{SnackbarContent}
\title{SnackbarContent}
\usage{
SnackbarContent(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/snackbar-content/}
}
\details{
\itemize{
\item action \code{node} \cr Default is - The action to display. It renders after the message, at the end of the snackbar.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item message \code{node} \cr Default is - The message to display.
\item role \code{string} \cr Default is 'alert' The ARIA role attribute of the element.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
