% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{SpeedDial}
\alias{SpeedDial}
\title{SpeedDial}
\usage{
SpeedDial(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/speed-dial/}
}
\details{
\itemize{
\item ariaLabel \code{string} \cr Default is - The aria-label of the button element. Also used to provide the id for the SpeedDial element and its children.
\item children \code{node} \cr Default is - SpeedDialActions to display when the SpeedDial is open.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item direction \code{'down'| 'left'| 'right'| 'up'} \cr Default is 'up' The direction the actions open relative to the floating action button.
\item FabProps \code{object} \cr Default is  Props applied to the Fab element.
\item hidden \code{bool} \cr Default is FALSE If true, the SpeedDial is hidden.
\item icon \code{node} \cr Default is - The icon to display in the SpeedDial Fab. The SpeedDialIcon component provides a default Icon with animation.
\item onClose \code{func} \cr Default is - Callback fired when the component requests to be closed.Signature:function(event: object, reason: string) =  voidevent The event source of the callback.reason Can be: "toggle", "blur", "mouseLeave", "escapeKeyDown".
\item onOpen \code{func} \cr Default is - Callback fired when the component requests to be open.Signature:function(event: object, reason: string) =  voidevent The event source of the callback.reason Can be: "toggle", "focus", "mouseEnter".
\item open \code{bool} \cr Default is - If true, the component is shown.
\item openIcon \code{node} \cr Default is - The icon to display in the SpeedDial Fab when the SpeedDial is open.
\item slotProps \verb{\{ root?: func| object, transition?: func| object \}} \cr Default is  The props used for each slot inside.
\item slots \verb{\{ root?: elementType, transition?: elementType \}} \cr Default is  The components used for each slot inside.
\item sx \code{Array} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item TransitionComponent \code{elementType} \cr Default is Zoom * @deprecated Use \code{slots.transition} instead. This prop will be removed in a future major release. \href{/material-ui/migration/migrating-from-deprecated-apis/}{How to migrate} The component used for the transition. Follow this guide to learn more about the requirements for this component.Deprecated Use slots.transition instead. This prop will be removed in a future major release. How to migrate
\item transitionDuration \verb{number| \{ appear?: number, enter?: number, exit?: number \}} \cr Default is    enter: theme.transitions.duration.enteringScreen,   exit: theme.transitions.duration.leavingScreen,  The duration for the transition, in milliseconds. You may specify a single timeout for all transitions, or individually with an object.
\item TransitionProps \code{object} \cr Default is - Props applied to the transition element. By default, the element is based on this Transition component.Deprecated Use slotProps.transition instead. This prop will be removed in a future major release. How to migrate
}
}
