% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{Step}
\alias{Step}
\title{Step}
\usage{
Step(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/step/}
}
\details{
\itemize{
\item active \code{bool} \cr Default is - Sets the step as active. Is passed to child components.
\item children \code{node} \cr Default is - Should be Step sub-components such as StepLabel, StepContent.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item completed \code{bool} \cr Default is - Mark the step as completed. Is passed to child components.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item disabled \code{bool} \cr Default is - If true, the step is disabled, will also disable the button if StepButton is a child of Step. Is passed to child components.
\item expanded \code{bool} \cr Default is FALSE Expand the step.
\item index \code{integer} \cr Default is - The position of the step. The prop defaults to the value inherited from the parent Stepper component.
\item last \code{bool} \cr Default is - If true, the Step is displayed as rendered last. The prop defaults to the value inherited from the parent Stepper component.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
