% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/inputs.R
\name{StepButton}
\alias{StepButton}
\alias{StepButton.shinyInput}
\alias{updateStepButton.shinyInput}
\title{StepButton}
\usage{
StepButton(...)

StepButton.shinyInput(inputId, ...)

updateStepButton.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to the component.}

\item{inputId}{ID of the component.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/step-button/}
}
\details{
\itemize{
\item children \code{node} \cr Default is NA Can be a StepLabel or a node to place inside StepLabel as children.
\item classes \code{object} \cr Default is NA Override or extend the styles applied to the component.See CSS classes API below for more details.
\item icon \code{node} \cr Default is NA The icon displayed by the step label.
\item optional \code{node} \cr Default is NA The optional node to display.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is NA The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
