% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{StepIcon}
\alias{StepIcon}
\title{StepIcon}
\usage{
StepIcon(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/step-icon/}
}
\details{
\itemize{
\item active \code{bool} \cr Default is FALSE Whether this step is active.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item completed \code{bool} \cr Default is FALSE Mark the step as completed. Is passed to child components.
\item error \code{bool} \cr Default is FALSE If true, the step is marked as failed.
\item icon \code{node} \cr Default is - The label displayed in the step icon.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
