% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{StepLabel}
\alias{StepLabel}
\title{StepLabel}
\usage{
StepLabel(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/step-label/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - In most cases will simply be a string containing a title for the label.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item componentsProps \verb{\{ label?: object \}} \cr Default is The props used for each slot inside.Deprecated use the slotProps prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item error \code{bool} \cr Default is FALSE If true, the step is marked as failed.
\item icon \code{node} \cr Default is - Override the default label of the step icon.
\item optional \code{node} \cr Default is - The optional node to display.
\item slotProps \verb{\{ label?: func| object, root?: func| object, stepIcon?: func| object \}} \cr Default is The props used for each slot inside.
\item slots \verb{\{ label?: elementType, root?: elementType, stepIcon?: elementType \}} \cr Default is The components used for each slot inside.
\item StepIconComponent \code{elementType} \cr Default is - The component to render in place of the StepIcon.Deprecated Use slots.stepIcon instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item StepIconProps \code{object} \cr Default is - Props applied to the StepIcon element.Deprecated Use slotProps.stepIcon instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
