% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Table}
\alias{Table}
\title{Table}
\usage{
Table(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/table/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the table, normally TableHead and TableBody.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item padding \code{'checkbox'| 'none'| 'normal'} \cr Default is 'normal' Allows TableCells to inherit padding of the Table.
\item size \code{'medium'| 'small'| string} \cr Default is 'medium' Allows TableCells to inherit size of the Table.
\item stickyHeader \code{bool} \cr Default is FALSE Set the header sticky.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
\examples{
data <- head(mtcars[, 1:3])
TableContainer(
  Table(
    size = "small",
    TableHead(TableRow(
      TableCell("Car"),
      lapply(names(data), function(col) TableCell(align = "right", col))
    )),
    TableBody(lapply(seq_len(nrow(data)), function(i) {
      TableRow(
        TableCell(component = "th", scope = "row", rownames(data)[i]),
        lapply(names(data), function(col) TableCell(align = "right", data[[col]][i]))
      )
    }))
  )
)
}
