% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{TableCell}
\alias{TableCell}
\title{TableCell}
\usage{
TableCell(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/table-cell/}
}
\details{
\itemize{
\item align \code{'center'| 'inherit'| 'justify'| 'left'| 'right'} \cr Default is 'inherit' Set the text-align on the table cell content.Monetary or generally number fields should be right aligned as that allows you to add them up quickly in your head without having to worry about decimals.
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item padding \code{'checkbox'| 'none'| 'normal'} \cr Default is - Sets the padding applied to the cell. The prop defaults to the value ('default') inherited from the parent Table component.
\item scope \code{string} \cr Default is - Set scope attribute.
\item size \code{'medium'| 'small'| string} \cr Default is - Specify the size of the cell. The prop defaults to the value ('medium') inherited from the parent Table component.
\item sortDirection \code{'asc'| 'desc'| false} \cr Default is - Set aria-sort direction.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \code{'body'| 'footer'| 'head'| string} \cr Default is - Specify the cell type. The prop defaults to the value inherited from the parent TableHead, TableBody, or TableFooter components.
}
}
