% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{TableContainer}
\alias{TableContainer}
\title{TableContainer}
\usage{
TableContainer(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/table-container/}
}
\details{
\itemize{
\item children \code{node} \cr Default is NA The content of the component, normally Table.
\item classes \code{object} \cr Default is NA Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is NA The component used for the root node. Either a string to use a HTML element or a component.
\item sx \verb{Array func| object| bool | func| object} \cr Default is NA The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
