% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{TableRow}
\alias{TableRow}
\title{TableRow}
\usage{
TableRow(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/table-row/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - Should be valid  tr  children such as TableCell.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item hover \code{bool} \cr Default is FALSE If true, the table row will shade on hover.
\item selected \code{bool} \cr Default is FALSE If true, the table row will have the selected shading.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
