% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{TableSortLabel}
\alias{TableSortLabel}
\title{TableSortLabel}
\usage{
TableSortLabel(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/table-sort-label/}
}
\details{
\itemize{
\item active \code{bool} \cr Default is FALSE If true, the label will have the active styling (should be true for the sorted column).
\item children \code{node} \cr Default is - Label contents, the arrow will be appended automatically.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item direction \code{'asc'| 'desc'} \cr Default is 'asc' The current sort direction.
\item hideSortIcon \code{bool} \cr Default is FALSE Hide sort icon when active is false.
\item IconComponent \code{elementType} \cr Default is ArrowDownwardIcon Sort icon to use.
\item slotProps \verb{\{ icon?: func| object, root?: func| object \}} \cr Default is The props used for each slot inside.
\item slots \verb{\{ icon?: elementType, root?: elementType \}} \cr Default is The components used for each slot inside.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
