% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/inputs.R
\name{TabList}
\alias{TabList}
\alias{TabList.shinyInput}
\alias{updateTabList.shinyInput}
\title{TabList}
\usage{
TabList(...)

TabList.shinyInput(inputId, ..., value = defaultValue)

updateTabList.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to the component.}

\item{inputId}{ID of the component.}

\item{value}{Starting value.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/tab-list/}
}
\details{
\itemize{
\item children \code{node} \cr Default is NA A list of Tab / elements.
}
}
