% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/inputs.R
\name{TabPanel}
\alias{TabPanel}
\alias{TabPanel.static}
\alias{TabPanel.shinyInput}
\alias{updateTabPanel.shinyInput}
\title{TabPanel}
\usage{
TabPanel(...)

TabPanel.static(...)

TabPanel.shinyInput(inputId, ..., value = defaultValue)

updateTabPanel.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to the component.}

\item{inputId}{ID of the component.}

\item{value}{Starting value.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
`TabPanel.static()` is an alias for `TabPanel()`, kept for backward compatibility.

\url{https://mui.com/material-ui/api/tab-panel/}
}
\details{
\itemize{
\item value \code{number| string} \cr Default is - The value of the corresponding Tab. Must use the index of the Tab when no value was passed to Tab.
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item keepMounted \code{bool} \cr Default is FALSE Always keep the children in the DOM.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
\note{
\code{TabPanel} is part of \href{https://mui.com/material-ui/about-the-lab/}{\code{@mui/lab}},
  which is published on the MUI beta channel. Lab APIs may change in future minor
  releases.

\code{TabPanel} is a display-only component: it shows or hides its
  content based on the active tab value held by the surrounding
  \code{TabContext}, but never fires \code{onChange} from user interaction.
  As a result \code{input$<inputId>} will always remain \code{NULL}. To
  react to tab changes on the server, read \code{input$<inputId>} from the
  \code{TabList.shinyInput} instead.
}
