% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{TextareaAutosize}
\alias{TextareaAutosize}
\title{TextareaAutosize}
\usage{
TextareaAutosize(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/textarea-autosize/}
}
\details{
\itemize{
\item maxRows \code{number| string} \cr Default is - Maximum number of rows to display.
\item minRows \code{number| string} \cr Default is 1 Minimum number of rows to display.
}
}
