% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Timeline}
\alias{Timeline}
\title{Timeline}
\usage{
Timeline(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/timeline/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item className \code{string} \cr Default is - className applied to the root element.
\item position \code{'alternate-reverse'| 'alternate'| 'left'| 'right'} \cr Default is 'right' The position where the TimelineContent should appear relative to the time axis.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
\note{
\code{Timeline} and its sub-components (\code{TimelineItem}, \code{TimelineDot},
  etc.) are part of \href{https://mui.com/material-ui/about-the-lab/}{\code{@mui/lab}},
  which is published on the MUI beta channel. Lab APIs may change in future minor
  releases.
}
\examples{
Timeline(
  TimelineItem(TimelineSeparator(TimelineDot(), TimelineConnector()), TimelineContent("Eat")),
  TimelineItem(TimelineSeparator(TimelineDot()), TimelineContent("Sleep"))
)
}
