% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Tooltip}
\alias{Tooltip}
\title{Tooltip}
\usage{
Tooltip(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/tooltip/}
}
\details{
\itemize{
\item children \code{element} \cr Default is - Tooltip reference element.This needs to be able to hold a ref.
\item arrow \code{bool} \cr Default is FALSE If true, adds an arrow to the tooltip.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item components \verb{\{ Arrow?: elementType, Popper?: elementType, Tooltip?: elementType, Transition?: elementType \}} \cr Default is The components used for each slot inside.Deprecated use the slots prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item componentsProps \verb{\{ arrow?: object, popper?: object, tooltip?: object, transition?: object \}} \cr Default is The extra props for the slot components. You can override the existing props or add new ones.Deprecated use the slotProps prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item describeChild \code{bool} \cr Default is FALSE Set to true if the title acts as an accessible description. By default the title acts as an accessible label for the child.
\item disableFocusListener \code{bool} \cr Default is FALSE Do not respond to focus-visible events.
\item disableHoverListener \code{bool} \cr Default is FALSE Do not respond to hover events.
\item disableInteractive \code{bool} \cr Default is FALSE Makes a tooltip not interactive, i.e. it will close when the user hovers over the tooltip before the leaveDelay is expired.
\item disableTouchListener \code{bool} \cr Default is FALSE Do not respond to long press touch events.
\item enterDelay \code{number} \cr Default is 100 The number of milliseconds to wait before showing the tooltip. This prop won't impact the enter touch delay (enterTouchDelay).
\item enterNextDelay \code{number} \cr Default is 0 The number of milliseconds to wait before showing the tooltip when one was already recently opened.
\item enterTouchDelay \code{number} \cr Default is 700 The number of milliseconds a user must touch the element before showing the tooltip.
\item followCursor \code{bool} \cr Default is FALSE If true, the tooltip follow the cursor over the wrapped element.
\item id \code{string} \cr Default is - This prop is used to help implement the accessibility logic. If you don't provide this prop. It falls back to a randomly generated id.
\item leaveDelay \code{number} \cr Default is 0 The number of milliseconds to wait before hiding the tooltip. This prop won't impact the leave touch delay (leaveTouchDelay).
\item leaveTouchDelay \code{number} \cr Default is 1500 The number of milliseconds after the user stops touching an element before hiding the tooltip.
\item onClose \code{func} \cr Default is - Callback fired when the component requests to be closed.Signature:function(event: React.SyntheticEvent) = voidevent The event source of the callback.
\item onOpen \code{func} \cr Default is - Callback fired when the component requests to be open.Signature:function(event: React.SyntheticEvent) = voidevent The event source of the callback.
\item open \code{bool} \cr Default is - If true, the component is shown.
\item placement \code{'auto-end'| 'auto-start'| 'auto'| 'bottom-end'| 'bottom-start'| 'bottom'| 'left-end'| 'left-start'| 'left'| 'right-end'| 'right-start'| 'right'| 'top-end'| 'top-start'| 'top'} \cr Default is 'bottom' Tooltip placement.
\item PopperComponent \code{elementType} \cr Default is - The component used for the popper.Deprecated use the slots.popper prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item PopperProps \code{object} \cr Default is Props applied to the Popper element.Deprecated use the slotProps.popper prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item slotProps \verb{\{ arrow?: func| object, popper?: func| object, tooltip?: func| object, transition?: func| object \}} \cr Default is The props used for each slot inside.
\item slots \verb{\{ arrow?: elementType, popper?: elementType, tooltip?: elementType, transition?: elementType \}} \cr Default is The components used for each slot inside.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item title \code{node} \cr Default is - Tooltip title. Zero-length titles string, undefined, null and false are never displayed.
\item TransitionComponent \code{elementType} \cr Default is - The component used for the transition. Follow this guide to learn more about the requirements for this component.Deprecated use the slots.transition prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item TransitionProps \code{object} \cr Default is Props applied to the transition element. By default, the element is based on this Transition component.Deprecated use the slotProps.transition prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
}
}
\examples{
Tooltip(title = "Delete", Button("Hover me"))
}
