% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/examples.R
\name{Typography}
\alias{Typography}
\title{Typography}
\usage{
Typography(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/typography/}
}
\details{
\itemize{
\item align \code{'center'| 'inherit'| 'justify'| 'left'| 'right'} \cr Default is 'inherit' Set the text-align on the component.
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item color \code{'primary'| 'secondary'| 'success'| 'error'| 'info'| 'warning'| 'textPrimary'| 'textSecondary'| 'textDisabled'| string} \cr Default is - The color of the component. It supports both default and custom theme colors, which can be added as shown in the palette customization guide.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item gutterBottom \code{bool} \cr Default is FALSE If true, the text will have a bottom margin.
\item noWrap \code{bool} \cr Default is FALSE If true, the text will not wrap, but instead will truncate with a text overflow ellipsis.Note that text overflow can only happen with block or inline-block level elements (the element needs to have a width in order to overflow).
\item paragraph \code{bool} \cr Default is FALSE If true, the element will be a paragraph element.Deprecated Use the component prop instead. This prop will be removed in a future major release. See Migrating from deprecated APIs for more details.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \code{'body1'| 'body2'| 'button'| 'caption'| 'h1'| 'h2'| 'h3'| 'h4'| 'h5'| 'h6'| 'inherit'| 'overline'| 'subtitle1'| 'subtitle2'| string} \cr Default is 'body1' Applies the theme typography styles.
\item variantMapping \code{object} \cr Default is h1: 'h1', h2: 'h2', h3: 'h3', h4: 'h4', h5: 'h5', h6: 'h6', subtitle1: 'h6', subtitle2: 'h6', body1: 'p', body2: 'p', inherit: 'p', The component maps the variant prop to a range of different HTML element types. For instance, subtitle1 to h6 . If you wish to change that mapping, you can provide your own. Alternatively, you can use the component prop.
}
}
\examples{
Box(Typography(variant = "h1", "h1. Heading"), Typography(variant = "body1", "body1. Text"))
}
