% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{Zoom}
\alias{Zoom}
\title{Zoom}
\usage{
Zoom(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/zoom/}
}
\details{
\itemize{
\item children \code{element} \cr Default is - A single child content element.This needs to be able to hold a ref.
\item addEndListener \code{func} \cr Default is - Add a custom transition end trigger. Called with the transitioning DOM node and a done callback. Allows for more fine grained transition end logic. Note: Timeouts are still used as a fallback if provided.
\item appear \code{bool} \cr Default is TRUE Perform the enter transition when it first mounts if in is also true. Set this to false to disable this behavior.
\item easing \verb{\{ enter?: string, exit?: string \}| string} \cr Default is - The transition timing function. You may specify a single easing or a object containing enter and exit values.
\item in \code{bool} \cr Default is - If true, the component will transition in.
\item timeout \verb{number| \{ appear?: number, enter?: number, exit?: number \}} \cr Default is    enter: theme.transitions.duration.enteringScreen,   exit: theme.transitions.duration.leavingScreen,  The duration for the transition, in milliseconds. You may specify a single timeout for all transitions, or individually with an object.
}
}
