% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{Breadcrumbs}
\alias{Breadcrumbs}
\title{Breadcrumbs}
\usage{
Breadcrumbs(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/breadcrumbs/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item expandText \code{string} \cr Default is 'Show path' Override the default label for the expand button.For localization purposes, you can use the provided translations.
\item itemsAfterCollapse \code{integer} \cr Default is 1 If max items is exceeded, the number of items to show after the ellipsis.
\item itemsBeforeCollapse \code{integer} \cr Default is 1 If max items is exceeded, the number of items to show before the ellipsis.
\item maxItems \code{integer} \cr Default is 8 Specifies the maximum number of breadcrumbs to display. When there are more than the maximum number, only the first itemsBeforeCollapse and last itemsAfterCollapse will be shown, with an ellipsis in between.
\item separator \code{node} \cr Default is '/' Custom separator node.
\item slotProps \verb{\{ collapsedIcon?: func| object \}} \cr Default is  The props used for each slot inside the Breadcumb.
\item slots \verb{\{ CollapsedIcon?: elementType \}} \cr Default is  The components used for each slot inside the Breadcumb. Either a string to use a HTML element or a component.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
