% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{ButtonGroup}
\alias{ButtonGroup}
\title{ButtonGroup}
\usage{
ButtonGroup(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/button-group/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item color \verb{'inherit'| 'primary'| 'secondary'| 'error'| 'info'| 'success'| 'warning'| string} \cr Default is 'primary' The color of the component. It supports both default and custom theme colors, which can be added as shown in the palette customization guide.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item disabled \code{bool} \cr Default is FALSE If true, the component is disabled.
\item disableElevation \code{bool} \cr Default is FALSE If true, no elevation is used.
\item disableFocusRipple \code{bool} \cr Default is FALSE If true, the button keyboard focus ripple is disabled.
\item disableRipple \code{bool} \cr Default is FALSE If true, the button ripple effect is disabled.
\item fullWidth \code{bool} \cr Default is FALSE If true, the buttons will take up the full width of its container.
\item orientation \verb{'horizontal'| 'vertical'} \cr Default is 'horizontal' The component orientation (layout flow direction).
\item size \verb{'small'| 'medium'| 'large'| string} \cr Default is 'medium' The size of the component. small is equivalent to the dense button styling.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \verb{'contained'| 'outlined'| 'text'| string} \cr Default is 'outlined' The variant to use.
}
}
