% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggerId.R
\name{Drawer.triggerId}
\alias{Drawer.triggerId}
\title{Drawer.triggerId}
\usage{
Drawer.triggerId(triggerId, ...)
}
\arguments{
\item{triggerId}{HTML id of an existing DOM element that acts as the trigger (button, link, etc.) to open the Drawer.}

\item{...}{args to pass to element}
}
\value{
Object with `shiny.tag` class suitable for use in the UI of a Shiny app.
}
\description{
Custom Drawer for mobiles using Drawer component. See 'js/src/Drawer.triggerId.jsx'.
}
