% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{Link}
\alias{Link}
\title{Link}
\usage{
Link(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/link/}
}
\details{
\itemize{
\item children \code{node} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item color \code{'primary'| 'secondary'| 'success'| 'error'| 'info'| 'warning'| 'textPrimary'| 'textSecondary'| 'textDisabled'| string} \cr Default is 'primary' The color of the link.
\item component \verb{element type} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.This needs to be able to hold a ref.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item TypographyClasses \code{object} \cr Default is - classes prop applied to the Typography element.
\item underline \code{'always'| 'hover'| 'none'} \cr Default is 'always' Controls when the link should have an underline.
\item variant \code{'body1'| 'body2'| 'button'| 'caption'| 'h1'| 'h2'| 'h3'| 'h4'| 'h5'| 'h6'| 'inherit'| 'overline'| 'subtitle1'| 'subtitle2'| string} \cr Default is 'inherit' Applies the theme typography styles.
}
}
