% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-extra.R, R/inputs.R
\name{LoadingButton}
\alias{LoadingButton}
\alias{LoadingButton.shinyInput}
\alias{updateLoadingButton.shinyInput}
\title{LoadingButton}
\usage{
LoadingButton(...)

LoadingButton.shinyInput(inputId, ...)

updateLoadingButton.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to the component.}

\item{inputId}{ID of the component.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/loading-button/}
}
