% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggerId.R
\name{Menu.triggerId}
\alias{Menu.triggerId}
\title{Menu.triggerId}
\usage{
Menu.triggerId(triggerId, ...)
}
\arguments{
\item{triggerId}{HTML id of an existing DOM element that acts as the trigger (button, link, etc.) to open the Menu.}

\item{...}{args to pass to element}
}
\value{
Object with `shiny.tag` class suitable for use in the UI of a Shiny app.
}
\description{
Custom Menu using Menu component. See 'js/src/MuiMenuTriggerId.jsx'.
}
