% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{MenuList}
\alias{MenuList}
\title{MenuList}
\usage{
MenuList(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/menu-list/}
}
\details{
\itemize{
\item autoFocus \code{bool} \cr Default is FALSE If true, will focus the role="menu" container and move into tab order.
\item autoFocusItem \code{bool} \cr Default is FALSE If true, will focus the first menuitem if variant="menu" or selected item if variant="selectedMenu".
\item children \code{node} \cr Default is - MenuList contents, normally MenuItems.
\item disabledItemsFocusable \code{bool} \cr Default is FALSE If true, will allow focus on disabled items.
\item disableListWrap \code{bool} \cr Default is FALSE If true, the menu items will not wrap focus.
\item variant \verb{'menu'| 'selectedMenu'} \cr Default is 'selectedMenu' The variant to use. Use menu to prevent selected items from impacting the initial focus and the vertical alignment relative to the anchor element.
}
}
