% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggerId.R
\name{Modal.triggerId}
\alias{Modal.triggerId}
\title{Modal.triggerId}
\usage{
Modal.triggerId(triggerId, ...)
}
\arguments{
\item{triggerId}{HTML id of an existing DOM element that acts as the trigger to open the Modal.}

\item{...}{args to pass to element}
}
\value{
Object with `shiny.tag` class suitable for use in the UI of a Shiny app.
}
\description{
Custom Modal bound to a DOM element by id. See 'js/src/MuiModalTriggerId.jsx'.
Open/close state is managed entirely client-side.
}
