% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{muiMaterialDependency}
\alias{muiMaterialDependency}
\title{Mui Charts UI JS dependency}
\usage{
muiMaterialDependency()
}
\value{
HTML dependency object.
}
\description{
Mui Charts UI JS dependency
}
