% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muiMaterialPage.R
\name{muiMaterialPage}
\alias{muiMaterialPage}
\title{HTML body without Bootstrap and margins}
\usage{
muiMaterialPage(
  ...,
  useFontRoboto = FALSE,
  useMaterialIconsFilled = FALSE,
  useMaterialIconsOutlined = FALSE,
  useMaterialIconsRounded = FALSE,
  useMaterialIconsTwoTones = FALSE,
  suppressBootstrap = TRUE,
  styleBody = "margin:0",
  debugReact = FALSE
)
}
\arguments{
\item{...}{The contents of the document body.}

\item{useFontRoboto}{Use Google Roboto font CDN in head, FALSE by default.}

\item{useMaterialIconsFilled}{Use Google icons CDN in head to use `Icon()` component, FALSE by default.}

\item{useMaterialIconsOutlined}{Use Google icons CDN in head to use `Icon()` component, FALSE by default.}

\item{useMaterialIconsRounded}{Use Google icons CDN in head to use `Icon()` component, FALSE by default.}

\item{useMaterialIconsTwoTones}{Use Google icons CDN in head to use `Icon()` component, FALSE by default.}

\item{suppressBootstrap}{Whether to suppress Bootstrap.}

\item{styleBody}{CSS style in body, using `margin:0` by default.}

\item{debugReact}{Whether to enable react debug mode. Default to FALSE.}
}
\value{
html object with 'margin:0' which can be passed as the UI of a Shiny app.
}
\description{
Creates a Material UI page without Bootstrap and with 0 margin in body by default.
You can choose to use Google Roboto font as well as Google icons fonts
with the `Icon()` component.
}
\details{
The Bootstrap library is suppressed by default, as it doesn't work well
with Material UI in general.


\url{https://fonts.google.com/icons?icon.set=Material+Icons}
}
