% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{Pagination}
\alias{Pagination}
\title{Pagination}
\usage{
Pagination(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/pagination/}
}
\details{
\itemize{
\item boundaryCount \code{integer} \cr Default is 1 Number of always visible pages at the beginning and end.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item color \code{'primary'| 'secondary'| 'standard'| string} \cr Default is 'standard' The active color. It supports both default and custom theme colors, which can be added as shown in the palette customization guide.
\item count \code{integer} \cr Default is 1 The total number of pages.
\item defaultPage \code{integer} \cr Default is 1 The page selected by default when the component is uncontrolled.
\item disabled \code{bool} \cr Default is FALSE If true, the component is disabled.
\item getItemAriaLabel \code{func} \cr Default is - Accepts a function which returns a string value that provides a user-friendly name for the current page. This is important for screen reader users.For localization purposes, you can use the provided translations.Signature:function(type: string, page: number | null, selected: boolean) =  stringtype The link or button type to format ('page' | 'first' | 'last' | 'next' | 'previous' | 'start-ellipsis' | 'end-ellipsis'). Defaults to 'page'.page The page number to format.selected If true, the current page is selected.
\item hideNextButton \code{bool} \cr Default is FALSE If true, hide the next-page button.
\item hidePrevButton \code{bool} \cr Default is FALSE If true, hide the previous-page button.
\item onChange \code{func} \cr Default is - Callback fired when the page is changed.Signature:function(event: React.ChangeEvent, page: number) =  voidevent The event source of the callback.page The page selected.
\item page \code{integer} \cr Default is - The current page. Unlike TablePagination, which starts numbering from 0, this pagination starts from 1.
\item renderItem \code{func} \cr Default is (item) =   PaginationItem ...item /  Render the item.Signature:function(params: PaginationRenderItemParams) =  ReactNodeparams The props to spread on a PaginationItem.
\item shape \code{'circular'| 'rounded'} \cr Default is 'circular' The shape of the pagination items.
\item showFirstButton \code{bool} \cr Default is FALSE If true, show the first-page button.
\item showLastButton \code{bool} \cr Default is FALSE If true, show the last-page button.
\item siblingCount \code{integer} \cr Default is 1 Number of always visible pages before and after the current page.
\item size \code{'small'| 'medium'| 'large'| string} \cr Default is 'medium' The size of the component.
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
\item variant \code{'outlined'| 'text'| string} \cr Default is 'text' The variant to use.
}
}
