% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggerId.R
\name{Popover.triggerId}
\alias{Popover.triggerId}
\title{Popover.triggerId}
\usage{
Popover.triggerId(triggerId, ...)
}
\arguments{
\item{triggerId}{HTML id of an existing DOM element that acts as the anchor/trigger for the Popover.}

\item{...}{args to pass to element}
}
\value{
Object with `shiny.tag` class suitable for use in the UI of a Shiny app.
}
\description{
Custom Popover bound to a DOM element by id. See 'js/src/MuiPopoverTriggerId.jsx'.
The trigger element acts as the anchor; the Popover opens on click and closes on clickaway.
}
