% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/inputs.R
\name{RadioGroup}
\alias{RadioGroup}
\alias{RadioGroup.shinyInput}
\alias{updateRadioGroup.shinyInput}
\title{RadioGroup}
\usage{
RadioGroup(...)

RadioGroup.shinyInput(inputId, ..., value = defaultValue)

updateRadioGroup.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to the component.}

\item{inputId}{ID of the component.}

\item{value}{Starting value.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/radio-group/}
}
\details{
\itemize{
\item children \code{node} \cr Default is NA The content of the component.
\item defaultValue \code{any} \cr Default is NA The default value. Use when the component is not controlled.
\item name \code{string} \cr Default is NA The name used to reference the value of the control. If you don't provide this prop, it falls back to a randomly generated name.
\item onChange \code{func} \cr Default is NA Callback fired when a radio button is selected.Signature:function(event: React.ChangeEvent, value: string) = voidevent The event source of the callback.value The value of the selected radio button. You can pull out the new value by accessing event.target.value (string).
\item value \code{any} \cr Default is NA Value of the selected radio button. The DOM API casts this to a string.
}
}
