% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{Stepper}
\alias{Stepper}
\title{Stepper}
\usage{
Stepper(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/stepper/}
}
\details{
\itemize{
\item activeStep \code{integer} \cr Default is 0 Set the active step (zero based index). Set to -1 to disable all the steps.
\item alternativeLabel \code{bool} \cr Default is FALSE If set to 'true' and orientation is horizontal, then the step label will be positioned under the icon.
\item children \code{node} \cr Default is - Two or more Step / components.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item component \code{elementType} \cr Default is - The component used for the root node. Either a string to use a HTML element or a component.
\item connector \code{element} \cr Default is StepConnector / An element to be placed between each step.
\item nonLinear \code{bool} \cr Default is FALSE If set the Stepper will not assist in controlling steps for linear flow.
\item orientation \code{'horizontal'| 'vertical'} \cr Default is 'horizontal' The component orientation (layout flow direction).
\item sx \verb{Array func| object| bool | func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
