% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R, R/inputs.R,
%   R/lab.R
\name{TabContext}
\alias{TabContext}
\alias{TabContext.shinyInput}
\alias{updateTabContext.shinyInput}
\title{TabContext}
\usage{
TabContext(...)

TabContext.shinyInput(inputId, ..., value = defaultValue)

updateTabContext.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)

TabContext(...)

TabContext.shinyInput(inputId, ..., value = defaultValue)

updateTabContext.shinyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{...}{Props to pass to the component.}

\item{inputId}{ID of the component.}

\item{value}{Starting value.}

\item{session}{Object passed as the `session` argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/material-ui/api/tab-context/}
}
\details{
\itemize{
\item value \verb{number| string} \cr Default is NA The value of the currently selected Tab.
\item children \code{node} \cr Default is NA The content of the component.
}
}
